requirements.txt
playwright==1.48.0
requests==2.32.3
python-dotenv==1.0.1
jpholiday==0.1.8
